add SynthesisEngineBase
