+++ voicevox_engine/synthesis_engine/__init__.py
@@ -1,9 +1,11 @@
 from .forwarder import Forwarder
 from .make_synthesis_engine import make_synthesis_engine
 from .synthesis_engine import SynthesisEngine
+from .synthesis_engine_base import SynthesisEngineBase
 
 __all__ = [
     "Forwarder",
     "make_synthesis_engine",
     "SynthesisEngine",
+    "SynthesisEngineBase",
 ]
