+++ voicevox_engine/synthesis_engine/synthesis_engine_base.py
@@ -1,0 +1,135 @@
+from typing import List
+
+from ..full_context_label import extract_full_context_label
+from ..model import AccentPhrase, AudioQuery, Mora
+from ..mora_list import openjtalk_mora2text
+
+
+def mora_to_text(mora: str) -> str:
+    if mora[-1:] in ["A", "I", "U", "E", "O"]:
+        # 無声化母音を小文字に
+        mora = mora[:-1] + mora[-1].lower()
+    if mora in openjtalk_mora2text:
+        return openjtalk_mora2text[mora]
+    else:
+        return mora
+
+
+class SynthesisEngineBase:
+    def replace_phoneme_length(
+        self, accent_phrases: List[AccentPhrase], speaker_id: int
+    ) -> List[AccentPhrase]:
+        """
+        accent_phrasesの母音・子音の長さを設定する
+        Parameters
+        ----------
+        accent_phrases : List[AccentPhrase]
+            アクセント句モデルのリスト
+        speaker_id : int
+            話者ID
+        Returns
+        -------
+        accent_phrases : List[AccentPhrase]
+            母音・子音の長さが設定されたアクセント句モデルのリスト
+        """
+        raise Exception('Unimplemented')
+
+    def replace_mora_pitch(
+        self, accent_phrases: List[AccentPhrase], speaker_id: int
+    ) -> List[AccentPhrase]:
+        """
+        accent_phrasesの音高(ピッチ)を設定する
+        Parameters
+        ----------
+        accent_phrases : List[AccentPhrase]
+            アクセント句モデルのリスト
+        speaker_id : int
+            話者ID
+        Returns
+        -------
+        accent_phrases : List[AccentPhrase]
+            音高(ピッチ)が設定されたアクセント句モデルのリスト
+        """
+        raise Exception('Unimplemented')
+
+    def replace_mora_data(
+        self,
+        accent_phrases: List[AccentPhrase],
+        speaker_id: int,
+    ) -> List[AccentPhrase]:
+        return self.replace_mora_pitch(
+            accent_phrases=self.replace_phoneme_length(
+                accent_phrases=accent_phrases,
+                speaker_id=speaker_id,
+            ),
+            speaker_id=speaker_id,
+        )
+
+    def create_accent_phrases(self, text: str, speaker_id: int) -> List[AccentPhrase]:
+        if len(text.strip()) == 0:
+            return []
+
+        utterance = extract_full_context_label(text)
+        if len(utterance.breath_groups) == 0:
+            return []
+
+        return self.replace_mora_data(
+            accent_phrases=[
+                AccentPhrase(
+                    moras=[
+                        Mora(
+                            text=mora_to_text(
+                                "".join([p.phoneme for p in mora.phonemes])
+                            ),
+                            consonant=(
+                                mora.consonant.phoneme
+                                if mora.consonant is not None
+                                else None
+                            ),
+                            consonant_length=0 if mora.consonant is not None else None,
+                            vowel=mora.vowel.phoneme,
+                            vowel_length=0,
+                            pitch=0,
+                        )
+                        for mora in accent_phrase.moras
+                    ],
+                    accent=accent_phrase.accent,
+                    pause_mora=(
+                        Mora(
+                            text="、",
+                            consonant=None,
+                            consonant_length=None,
+                            vowel="pau",
+                            vowel_length=0,
+                            pitch=0,
+                        )
+                        if (
+                            i_accent_phrase == len(breath_group.accent_phrases) - 1
+                            and i_breath_group != len(utterance.breath_groups) - 1
+                        )
+                        else None
+                    ),
+                )
+                for i_breath_group, breath_group in enumerate(utterance.breath_groups)
+                for i_accent_phrase, accent_phrase in enumerate(
+                    breath_group.accent_phrases
+                )
+            ],
+            speaker_id=speaker_id,
+        )
+
+    def synthesis(self, query: AudioQuery, speaker_id: int):
+        """
+        音声合成クエリから音声合成に必要な情報を構成し、実際に音声合成を行う
+        Parameters
+        ----------
+        query : AudioQuery
+            音声合成クエリ
+        speaker_id : int
+            話者ID
+        Returns
+        -------
+        wave : numpy.ndarray
+            音声合成結果
+        """
+        raise Exception('Unimplemented')
